requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-telegram-bot[asyncio]==20.7
google-cloud-storage==2.14.0
google-auth==2.29.0
pyjwt==2.8.0
pydantic==2.5.3
pytest==7.4.4
pytest-asyncio==0.21.1
